feat : add filters on crm and partner and add association prospects

+++ foodeals-admin-api/src/main/java/net/foodeals/crm/application/dto/responses/ProspectStatisticDto.java
@@ -1,4 +1,4 @@
 package net.foodeals.crm.application.dto.responses;
 
-public record ProspectStatisticDto(Integer activeLeads, Long total, Integer notConverted, Integer converted) {
+public record ProspectStatisticDto(Long activeLeads, Long total, Long notConverted, Long converted) {
 }
